+++ geotrek/common/sql/10_utilities.sql
@@ -31,14 +31,14 @@ $$ LANGUAGE plpgsql;
 
 SELECT create_schema_if_not_exist('geotrek');
 
-CREATE OR REPLACE FUNCTION geotrek.ft_date_insert() RETURNS trigger AS $$
+CREATE OR REPLACE FUNCTION geotrek.ft_date_insert() RETURNS trigger SECURITY DEFINER AS $$
 BEGIN
     NEW.date_insert := statement_timestamp();
     RETURN NEW;
 END;
 $$ LANGUAGE plpgsql;
 
-CREATE OR REPLACE FUNCTION geotrek.ft_date_update() RETURNS trigger AS $$
+CREATE OR REPLACE FUNCTION geotrek.ft_date_update() RETURNS trigger SECURITY DEFINER AS $$
 BEGIN
     NEW.date_update := statement_timestamp();
     RETURN NEW;